requirements.txt
flask
python-dotenv
openai
requests
python-dateutil
isodate
gunicorn